All tests pass; Refactors; Bugfixes introduced from lazy saving;

+++ Application/Handlers/Products/Delete/DeleteProductHandler.cs
@@ -10,12 +10,14 @@ namespace Application.Handlers.Products.Delete;
 public class DeleteProductHandler : IRequestHandler<DeleteProductCommand, OneOf<DeleteProductResult, List<ApplicationError>>>
 {
     private readonly IProductDomainService _productDomainService;
+    private readonly IProductHistoryDomainService _productHistoryDomainService;
     private readonly IUnitOfWork _unitOfWork;
 
-    public DeleteProductHandler(IProductDomainService productDomainService, IUnitOfWork unitOfWork)
+    public DeleteProductHandler(IProductDomainService productDomainService, IUnitOfWork unitOfWork, IProductHistoryDomainService productHistoryDomainService)
     {
         _productDomainService = productDomainService;
         _unitOfWork = unitOfWork;
+        _productHistoryDomainService = productHistoryDomainService;
     }
 
     public async Task<OneOf<DeleteProductResult, List<ApplicationError>>> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
@@ -27,6 +29,9 @@ public class DeleteProductHandler : IRequestHandler<DeleteProductCommand, OneOf<
 
         var canDelete = await _productDomainService.TryOrchestrateDeleteProduct(product);
         if (canDelete.IsT1) return new CannotDeleteProductError(message: canDelete.AsT1, path: []).AsList();
+
+        var canInvalidate = await _productHistoryDomainService.InvalidateHistoryForProduct(product);
+        if (canInvalidate.IsT1) return new CannotInvalidateProductHistoryError(message: canInvalidate.AsT1, path: []).AsList();
 
         await _unitOfWork.SaveAsync();
         return new DeleteProductResult();
